Tokenizer and dataset change
Moving away from the 'plain text wikipedia' dataset and using a preprocessed dataset like wikitext-2-v1. Creating a new Tokenizer (one thats similar to GPT-2's tokenizer) which uses BPE rather than the current word level tokenization process. I tried to train a model using the old tokenizer (word level) and the train time was too long (so was the vocab-size), which lead to the change in dataset used and the tokenizer.