Built tokenizer
Built and saved a 'fitted' tokenizer on the training data. Will change the tokenizer's fit-data later for better generalization. Saved the tokenizer's _vocab and _merges into a pickle file, use the load_bpe_tokenizer function to load the tokenizer (for now, will update the class to load correctly). Will convert all of this into a Python script to make it easier to save,load and fit a tokenizer.